All in with Typescript and add tests
We go all in with Typescript, moving all sources to a `src` directory,
and add tests for the new `cmd.ts`.  We remove the use of `standard`
since we no longer have pure Javascript sources.

We start adding more tests to increase the coverage of `client.ts`.

We are dropping support for Windows since we don't have an equivalent
script to generate the version number.